fix: calculate earnings and amount type

+++ src/Entity/TotalEarnings.php
@@ -23,7 +23,7 @@ class TotalEarnings
     private $user;
 
     /**
-     * @ORM\Column(type="integer")
+     * @ORM\Column(type="float")
      */
     private $amount;
 
@@ -49,12 +49,12 @@ class TotalEarnings
         return $this;
     }
 
-    public function getAmount(): ?int
+    public function getAmount(): ?float
     {
         return $this->amount;
     }
 
-    public function setAmount(int $amount): self
+    public function setAmount(float $amount): self
     {
         $this->amount = $amount;
